+++ src/to_curl.rs
@@ -444,7 +444,7 @@ pub fn translate(args: Cli) -> Result<Command> {
                 cmd.arg(json_string);
             }
             Body::Json(..) if args.json => {
-                cmd.header("content-type", JSON_CONTENT_TYPE);
+                // Don't set Content-Type: there's no JSON body to describe
                 cmd.header("accept", JSON_ACCEPT);
             }
             Body::Json(..) => {}
@@ -529,7 +529,7 @@ mod tests {
             ),
             (
                 "xh --json httpbin.org/post",
-                "curl http://httpbin.org/post -H 'content-type: application/json' -H 'accept: application/json, */*;q=0.5'",
+                "curl http://httpbin.org/post -H 'accept: application/json, */*;q=0.5'",
             ),
             (
                 "xh --form httpbin.org/post x@/dev/null",